requirements.txt
requests==2.20.1
GitPython==2.1.11
